Return a List of Conferences.
When asking for all the Conferences of a given resource type, it doesn't
make sense to ask for only one, or None (and throw an exception). It
seems more sensible to request the whole list, or even a paginated list.
Since this method isn't used anywhere in the code yet, we'll
just return all().