Enhancement/participant role filters (#1540)
This feature allows for more granular participant role resolution. Historically, Dispatch allowed you to manage which service would be resolve on a per-incident type basis. This meant that only one team would be assigned a role for a given incident (e.g. incident commander or liaison).

With this change, we create role policies that invert this relationship. Allowing the user to define aspects of an incident and then which team/service should be assigned a given role.

The typical use-case is for when a given team is capable of handling a low/medium priority incident but requires another team (e.g. a crisis management team) for high priority incidents.

When policies conflict with each other, it deconflicts by using the order of the policies as a measure of relative importance.