Update slack plugin.py to check if user exists before send direct
It is possible that a message could be attempted to send for a slack account that does not exist anymore. This performs a check before attempting to send so it does not raise an exception if the user does not exist.

This causes two looks ups for every direct or ephemeral send, but will prevent exceptions.