feat(slack): Expose create case slash command
In the slack conversation plugin, there are slash commands to be used on
cases after they are created, but there is no slash command for creating
a case. As I looked into this issue, I found that the full setup for
allowing the command is present, but the final wire-up of the command
was not. I added that final wire-up. I looked for tests for any
functionality like this and found none. I can create it, but I have no
model to go on.