feat(slack): Expose create case slash command (#5295)
* feat(slack): Expose create case slash command

In the slack conversation plugin, there are slash commands to be used on
cases after they are created, but there is no slash command for creating
a case. As I looked into this issue, I found that the full setup for
allowing the command is present, but the final wire-up of the command
was not. I added that final wire-up. I looked for tests for any
functionality like this and found none. I can create it, but I have no
model to go on.

* Copy previous plugin extension PR

I am copying work from https://github.com/Netflix/dispatch/pull/5240/files
to try and emulate better what I should do. I am pretty stuck on
generating an alembic file correctly for this type definition. I'll keep
trying to figure it out from the docs, but it's not transparent to me.

* add missing div close

* Fixes modal opening.

---------

Co-authored-by: Avery <averyl@netflix.com>