fix(slack): Make slack /events respond to challenge
When configuring event listening for a slack app, your endpoint is
required to respond correctly to the slack challenge that is sent. If
your endpoint does not respond correctly, slack will not let you use it.
Currently, the events endpoint does not respond correctly to the
challenge because it is using a background thread to process all
incoming events and responding with a static response. In order to use
the events endpoint correctly (including using an @ in the incident
channel to add an observer), that endpoint has to respond. I altered the
event post to check the body for a "url_verification" type and, if it
exists, process the event synchronously in the same way the `actions`
endpoint does. Otherwise, process async like it always has.